Updating struct GameWorld after comparing with book
the `snake` should be a `deque`, instead of a `vector`, we need
`pop_head` and `push_head` of it.

Comparing to the one in the book, my `GameWorld` doesn't have
`int numEaten` and `int numRows, numCols;`.

1. `numEaten` can be got easily with `snake.size()`.
2. In the book `numRows` and `numCols` are used to determine whether
   the snake is hitting the boundaries, but I think since the wall and
   boundaries share same character, e.g "*", so the code for determining
   wall and boundaries can be shared too, determining if hitting the
   wall is applicable for whether hitting the boundaries.

May be I'm wrong, if I am, I will return here and explain why.